fixed bug with remote annotation on types that have not been transformed
not sure why this just started showing up but we add the annotation even
for dynamic types in the 'transform' method. But I don't know if we should
be doing that since we don't have the equivalent of 'hidden modifiers' so this
could affect the source version of model if we look at it or save it.

instead just return the annotation on the dynamic type if there are remote runtimes
defined for this method