fix for goto class before layer has been opened
- sync fixes to allow type lookups to run without acquiring the dyn lock
- doInit flag to parse so that we can delay the initModel operation till after
the parse has been completed. this is for the intelliJ plugin so we can do that
entire thing unsynchronized to avoid contention and improve performance.

it turned out the major performance problem was due to a class loading that
intelliJ was doing where it seemed to lock up the JVM for a few seconds trying
to load the SCFileAction class (and probably all of the libraries used by the template
builder in intelliJ). but removing that lock did seem to speed up text editing a lot
and removes that deadlock problem with the psi lock.