fixed bug with ordering of dependencies in JS files
- this was kind of a major problem that just showed up when adding
a new dependency between two different type hierarchies. There was logic
to follow the dependencies of a type before following all of the extends
types when building the order for how types are put into the file. Only the
extends types form a DAG - the dependencies are a hairball so those need to
come after the extends tree.