fixed bug with incremental compiles
for some reason, this was a regression in the runIncrCompile test suite
when PageCompileServlet was changed. The PageInit.java file from jetty.schtml
file was being used for the downstream build layer as well and so missed TodoList.
When getting the previous type group members, need to use the previous transformed layer
if it's set... if we transform it once and it's not the same in next build layer, need
to transform it again.