separate flag for 'replacedInactive'
the replaced=true flag implies that replacedByType has the new type but for
inactive types we can no longer set that value. As a quick fix, adding
a new flag to store the fact that we have an inactive type which might be
stale... we can't refer to the new type due to some weird bug where it causes
a memory leak. Possibly someone has a ref to the first type and then all of the
rest stay in the heap?  We do need to know if a type has been replaced when
doing the refreshTypes process or resolving an extends type