fixed bug with valid status propagation
For a binding like:

  visible := editorModel.ctx.currentLayers.contains(typeLayer)

when 'typeLayer' changed from null to some other value which
is still not in currentLayers, the VariableBinding would get
marked with 'valid=false' when the event was invalidated but
would not get marked with 'valid=true' when it was validated.
We'd see that the method return value did not change and just
returned, rather than propagating that 'we are valid again'
message up to the parent binding (the VariableBinding). The
MethodBinding child then would not get invalidated the next
time currentLayers changed (because we stop propagating that
change once we see that valid=false on the parent binding)
It would then not get re-validated...