fixed bug with updateParseNodes
in a case where we had arguments and the
first entry had been converted from an 'expression' to a simple
string because of that optimization. It was in an array parselet
and the string was not matching the first element and so we failed
to find a match for the second and subsequent elements and kept
infinite looping - retrying the list each time.