fixed bug with replaceChild not using == for semantic node comparison
the code looked like:

        textArea.text = ctx.getModelText(model, model.modelText);

getting transformed into

        getTextArea().setText(getCtx().getModelText(getModel(), getModel().getModelText()))

but the two params were swapped.  In an intermediate stage, the two args were both 'getModel()' when
we called replaceChild and the wrong one was replaced. Later we changed that node to have the .getModelText()
so it mattered that we got the right instance, even if the nodes at the time were identical.