include List<String> as a non-reference object
set 'sync mode' so we are not applying changes when
resolving an object. this happens in a weird case - when the
client is committing changes to an object we have not yet
defined. So in resolving the object, we should be setting
properties to be sync'd to the server on init, but those were
seeing the sync state as 'applying changes' not 'recording changes'
so they were not sent.

non-reference properties can be serialized right away so we want
to put as many properties in that set as possible. For method return
values for example, the result handler can run before the properties
are set (unless you use the postListener). Ideally we should make
this configurable as a property option - maybe we have a way to just
mark a property as a 'first-pass' property or something like that
or have a way to override the default value in either direction...
we also should probably do this test using the declared type of the
property, not the current property value.