fixed bug with init sync children and sync action
for the client, when we initialize the children, assume we
are applying changes the server already has so property changes
made during the init process are not confused with those that
are already on the server. It's possible this will miss some
init code that only runs on the client but right now, we don't
have a way to reliably identify property changes which are made
on both client and server and so can exclude the initial
set for those properties.