request scope + move scoping of id/name generation to scopeContext
the static objectIds and typeIdCounts tables cause problems with
request scope for the todo sample which wants to have consistent
ids for each request. It makes more sense to scope the id spaces
anyway in the scope context where the object is defined when we generate
the ids from the type name at least.

fixes some other bugs in sync for request scope. The term reset was
being overloaded and misused in some method names so made a change
to fix that up. the client for request scope needs to send a reset
on each request - all but the init state.