put all update actions in the index and update newType
For the case where we have both UpdateType and UpdateProperty
in the same refresh, and there are stopped types processed which
causes an extra level of 'replace type' in the refresh (i.e.
the new type gets replaced again).

In this case, we need to update newType in all actions or else
we have a stale version of newType processed in the UpdateProperty.
Looked like the same bug applied to NewType as well.