fixed bug with interrupt layer restart and bug with activate layers
- for activate layers, we were clearing out the layers in the main
layered system inside of initRuntimes and then using those layers
to figure out which procs/runtimes we needed later in activateLayers.
Since we found no procs/runtimes once we cleared them out, we resorted
to the default runtime and cleared out all active layers in js/java_Server

- for the init layers bug, during an intelliJ JavaPsi.findClass operation
we were getting interrupted initializing a layer. the exception handling
happened after the layer model was put into the index though. So the
next time we tried to find that layer, we'd use this cached layer that
had not been initialized. changed the code so there's a new flag indicating
when a layer positively has been initialized. If we see that is not set
we don't just return the lyaer - instead, finish initializing the model
first.