fix for refresh codeView
- when applying a sync layer, we turn on queuing of binding events so
that we do not try to send events, call methods etc. on each property
change. But if a property change event triggers a change to a list
component which contains sync'd properties, we need to process that
list change during the applySync operation. This means flushing the
binding queue in order to sync the repeat tags (keeping the event in
order). We also wait until we try to resolve an element of the list
so we don't refresh too early.

- to make this most efficient, we don't want to flush the binding queue
every time we try to resolve a name for a list. We'd like to know that
the list value is possibly invalid, then flush the queue, and validate.
So to make that work, there's a new mode for the BindingContext - query_validate_only - which sends invalidate messages in immediate mode but
queues the validate ones.

- after putting in that change, it turns out the invalidate events
were not propagating all the way to the listeners on the property
itself so changed the binding code. whereas before it would not
send the invalidate event, and then on the vlaidate event it would
send value changed (i.e. both at once). Now it sends the events
independently after the first value-changed event is ent.