SyncTypeFilter annotation
eliminate use of globalSyncTypeNames

when we have some web pages that use js.layer and the layered  system
sync and others that do not, the global sync names were opening
up those sync objects to a process that did not have those type
definitions. So we improved the mechanism using an annotation that
lists other synchronized class names.