added setValueByRef to ScopeContext
this is for a case where the object is stored in another context
where it will be freed.

The specific case is that server tags are registered in the
window scope context but may live in app session scope context.

When the window context is destoryed, the server tag should not
be destroyed