post build commands, disabled runtimes/main system,
evalTemplate option to resolve in layer so that we
can use post build commands to do a 'sed' on the
transformed result in-place before the transform.

postBuild seems generally useful for integration as is
the option to eval templates in the context of the
layer definition files instead of just the object definition parameters
or other main object. For now, code-templates don't use
evalTemplate with the resolveInLayer flag to true but there might
be a use case for that down the road.