one-to-one fixes to DB stuff
- split up db template into the 'define types' part and the part that
defines on the properties to define the complete type descriptor.
  * We run the define types part during initTypeInfo now and it should be able to define any new interface methods using only the type, base-types, and list of properties (i.e. to identify the id properties). It could also later add query methods if we want.

unfortunately this version broke the many-to-many example because we need to further separate the DBTypeDescriptor init code so that we don't access any references to other DBTypeDescriptors until after initTypeInfo - otherwise, we end up needing to start those types before we've fully started this one.