comments and add errors for repeat and negated patterns in URLs
don't have a use case yet for those but nothing we couldn't implement
right now though we don't have a full pattern implementation on the
client side so maybe we'd move more of that code over. Right now, the
URL pattern gets converted to a list of strings/URL param property
objects that can manage going in both directions - i.e. building a URL
from an instance and updating the URL when those properties change