scsql extends sc
allows us to transform scsql into java with the same model classes
- copied some code from Template to make this happen. Either we could make
this more of a pattern that's easier to replicate in other formats, or
have SQLLanguage extend Template language. That might be useful anyway.

changed a lot of names to remove some conflicts created when we changed that
extension

order by - options for a fixed sent of ascending/descening properties or

db first - parse scsql files generated from the userAddressQuery example and
use them as the source files for userAddressQueryDBFirst.