fixes for sub-types, recursive references with eager properties
for the new inheritCatalog test that takes part of the ecommerce
model for parent/child categories, references to images

generalized the select query so that it can support the same table
joined more than once.

complete the lifecycle for initializing fetch groups - particularly
when there is a recursive relationship. We need to handle the references
differently - they get stored into a separate refProps list in the query
and then the references are resolved after the rest of the queries have
been put together.

added alter set data type for when a cast can be used and do drop/add column
otherwise