run's the first ec test - more DB refinements
- fixed probems with the case where a base class is used by more than
one persistent class in different table structures.  The tricky thing is
handling the persistent properties that are in a non-persistent base class.
I decided for example not to store Products and Categories in the same table
since they are always treated differently in the UI but there are a lot of
properties in common and they do share a base class.

This is handled now by adding a wrapper getX/setX that adds the persistence aspects.
I originally added a dbPropDesc to VariableDefinition in order to support pushing
persistence into the base class so that the getX/setX could be shared but that
seems less flexible. For now the dbPropDesc is mostly just a flag we use to make
sure the base class's property is converted to getX/setX from a field.

- override DBPropertySettings in a sub-type or layer

- sorting of sql commands - build a graph, do a breadth-first search on the graph
when determining the sorted list.