refId properties for single-valued references
This solves the problem when there is a single-valued association to
another type in a type hierarchy. We want to fetch the id of the other
type as a normal column of the referencing table in the first query.
Because we don't have the db_type_id we can't create a hollow instance for
that id and it's possible applications will want to use the id or test for
null without fetching the referenced object. So we add a propNameRefId property
in that case that is a public field in the generated IDBObject class
and set that during the first query. When the getX method on the propName is
called, it will then do a query for that id on the associated table picking up
db_type_id along with the other columns.

fixed bugs with the test id logic - store test ids by type