fixed bugs with explicit option tags
 - don't sync Option.selected because there's only one option instance
on the server and client for rendering all options. instead, we sync
Select.selectedIndex and that will change option.selected for each
instance, render the option tag, then go on to the next one with the
same instance.

It's possible that we should model option more like 'repeat' but so far
the rendering of option tags is pretty minimal so the performance loss
of making them basically stateless should not be great and this implementation is in some ways simpler