added api for a default value in the sync type
- right now, used only for server tag's liveEdit properties
but could be used by the code-generator so that we are more
consistent with how init is implemented

- *** initDefault at the class level no longer implies the 'init' flag
for the property if a SyncPropOption was specified. We need the ability
to turn on/off that flag and it seemed inconsistent that this flag was
only set for properties that had a SyncPropOption class.