CSV language, close prepared statement, import CSV, parselets tweaks
- fixed parselets bug where generating an array value into a ([],[]) where
the second one is scalar value. Before passing the list value into the first
one, need to trim off the last value from the list

- for generating a lineTerminator - symbol choice, there's a defaultGeneratedValue
so that it's not ambiguous

- support for SQLdata type with two words: "double precision"

- db fixes for floating point numbers

- importCSV utility - ended up not using the CSV parser but that might be useful
later on for read/write applications or where we are generating code directly from
the the CSV structure, to allow it to be modified or customized for example.  It also
found a useful bug and is a reasonable parser example because it's so small