db search query, live dyn types change for db types
- fixed bug with transform errors showing up with modified types

- we want liveDynTypes enabled on the server but only for persistent
objects (with an id) otherwise, there's no consistent id for them.
on the client, we want to just turn on liveDynTypes so you don't have
to set it. At some point, maybe we should have a way to turn it off
for production since we only want it enabled with the management ui.