fixed ordering bug with pendingNewObjects in sync layer
in formatExpression, when we encounter an object that's not yet initialized
we add call createOnDemandInst. But if it has been created but ends up
getting added to the sync system after the referencing inst, it might
be behind this instance in the syncChangesList and so won't be available
during serialization.

To fix this, keep track of all pendingNewObjects in the sync layer and remove
them as we serialize. If an object has a name, but is in the pendingNew for
formatExpression we can then just add the new object as a dependent change
and mark the old one as 'overridden' so it does not get serialized twice.