object id supported for transient objects
since object id is used for synchronization and logging, if we need
to sync a transient instance, we need a unique, consistent id for
that purpose. for now, tack a 't' onto the end of a sequence number
for that process. Ideally we'd have a way in the sync system to change
the id of an object so that once we make it persistent we could change
it but this will work as long as we don't need to sync that instance
against a new JVM.