fixed bindings that refer to the same object they are defined in
for tags if you use the tag id name, it would previously put a
getX method call to retrieve the object while in the binding init
code that caused a recursive infinite loop.  The code itself
should be probably using 'this' but this is an easy case to fix by just
replacing it with this in the generated code.