server side focus method should call the focus method on the client
implement the remote call through the page dispatcher so that we can
place it in the window scope used by the page. This seems like the
cleanest solution but turned out to be less important because we
end up queuing the focus call on the client anyway, so it doesn't matter
if it's in front of the innerHTML that defines the tag to be focused.