moved getPropertyType call for initDBTypeDescriptor
we only needed it for identity properties and since we might
be in the middle of initializing a type that we need to resolve
a property type this simplifies things a lot.  I think ultimately
we only really need the property type name but more code would
need to be rewritten for that and id properties are usually simple
types like long and string.