added scrollWidth/height, bug with server tags listeners
when listeners were added after the tag object was initialized (but before
the first render) those listened on properties were not added as
listenerProps on the client.

Now we add a listenerAdded callback for the server tag object and
invalidate the listener props and do a validateServerTags after
we have initialized all of them.