signature change for the matchQuery, searchQuery, and findBy
methods.  Put the list of properties up front, move selectGroup till
just before orderBy properties

allow searchQuery to use a list of properties to 'and' into the search
part of the query (that is formed by 'ORing ILIKE expressions right now)

when we search for a multi-valued property that has no rows, do populate
the List property with an empty list so that we know this value has been
fetched. Otherwise, the first time we set it we go and delete all rows
because we don't know we already fetched it.