removed code affecting resolve for inactive modifies
the question is how we choose the modify type for a given
in the inactive system. Previously there was logic to choose
the type explicitly inherited (if there was one) by the base
types of the ref layer. That seems like more of a workaround
for a problem where layers might not be sorted properly... or
there might be an incompatible intermediate layer in the way.

The tests all work with the new model and the IDE works much
better for the ecommerce bundle without that check. Just go
with the next layer in the stack for the modify type, like we do
when we compile the system.