added id parameter to Element constructor
for tag objects created by a RepeatWrapper... right now top-level
tags do not have the setId(allocUniqueId("xyz")) code put into them
because they are assigned an id from their parent. This sort of makes
sense since the id should be assigned in a context of the parent to
be sure it's unique in the page.

But, if there's no idea for these top-level tag objects, the server
tags support does not work - the getElementById call made in the client
fails cause there's no id.  The id should be assigned before we first
evaluate the startTagTxt so either in the constructor or as an early
instance field initializer. It seems like the most robust solution is
to pass the top-level id in the constructor for the repeated element so
that it's defined.