fixed bug with excluded type errors
if an inner type is excluded, errors for references in the
excluded process should not show up. needed to change the
start order, mark things as excluded, check if excluded parent
is before starting a reference

fixed bug with resolving active types where in strange cases the
most specific layer was not being returned. if we had cached
a type without setting last started we were not checking the
src file layer position.  this would show up when using the
dynamic system after starting the process

added support for client-side constructor properties that are
in the @URL annotation. in tags.js
now there is a method to resolve the value of a constructor
property and in CompilerSettings a property to specify a
constructor property resolver method that wraps the default value.

fixed bug with javascript where exec="server" on the top level
tag object was still generating a JS file for that type even
though the entire type was excluded. That is how you should specify
a server tags only page that's built with client/server layers