fixed event propagation for recordChange method
use childSyncInsts, not childContexts to find all multi-level child contexts
that might contain this item. we don't add the instance to all of the intermediate
SyncContexts so we won't find it with childContexts and it's a lot slower.  This
same change was made to the invalidate method but just missed this case for recordChange