fixed bug with refresh and on demand instances stored in a parent scope
When refreshing the page, the accessInst method needs to ensure that
we send definitions for any on-demand instances that are loaded into
that scope.

The other way I considered fixing this was to put the on-demand instance
into the initial sync layer of the app-session but with on-demand instances
the accessInst method will traverse the tree of accessible objects and
is the right way to determine which subset of app-session components belong
in the new window.