added repeatSync=true/false flag used primary for scripting
In the testEditBooks.scr file, we want to select an element of a
repeatTag which means that the repeatTags need to be synchronized
with the sync system. Previously, we would always register all
repeatTags as registered all the time which was slow and no longer
worked with the ProductView.childViews which just repeats over
the repeatTags themselves (i.e. repeat == repeatTags). and those
elements are synchronized via another property so doing this disabled
the 'new' record we need to send for those other cases.

It seems much better to "opt in" to this scripting behavior.  I even
tried the enableRepeatSync method but it was getting called too late
to be useful cause ideally we'd only want to turn this on for 'testMode'
when running the test script