bug fixes for dynamic mode testing
- db types for now compiled only. it would not be too much work to add this and some
interesting potential down the road but the main priority is getting templates to
refresh on top of the core db. unfortunately, this makes it hard to run the whole ecommerce
app in dynamic mode because once you set compiledOnly on types, more compiled code is
depending on dynamic types and that runs into bugs - i.e. you can't extend a dynamic type
from a compiled one right now reliably, and there are issues where a dynanic modify messes
up everything because it makes the compiled layers dynamic as well... kind of needs to generate
a partially compiled type and add dynamic behavior to that type I think - so only the modified
changes are in dynamic mode, new methods and fields but all compiled features in the lower level
layers remain compiled. I'm not sure if there is one correct behavior here - we might need something
to support both cases.

- dynCompiledMethods - need any unimplemented abstract/interface methods to be added to this list
so that the stub can be a concrete class, even if it's abstract. that way, we can avoid creating
extra stubs for each subclass

- set compiledOnly flag much earlier so we don't start making the type dynamic