fixed a few bugs with dynamic mode for running product demo
- need to run the init code for a type just before the constructor
so when invoking a constructor with the super(..) to another class
need to make sure the pending constructor's type is inited just after
it's super() call is made and before it's constructor is run.  There
were one or two cases where it was running the init code after the
constructor that are now fixed

- fixed regression with the change to not use 'checkExtends' in the
test before calling compileDynamicStub. We need to make sure the
extends type is compiled if necessary. The code for compileDynamicStub
now just returns if this type is not actually a dynamic stub. Also
fixed the bug for determining when a modified chain ends up hitting
a "compiled inner class" - i.e. one where the dynamic stub needs to
be part of the parent classes dynamic stub.

- fixed a problem resolving types caused by weird init order that happens
when running in incremental mode when there were no changes. In trying to
init a type, need to make sure we have called the initTypeInfo on the chain
before trying to get the DBTypeDescriptor. Otherwise, we end up initing the
type info for the base class - where we are generating the type info while
generating the type info.

- fixed invalid warnings for when using a compiled type for a property made
bindable due to being a persistent property

- fixed bug with "callsSuper" method where we were optionally not returning
the type for a modify type but not accurately checking if the super is a modify
super.