added resolveSrcType option for getting dependent types
not currently using this but it might be useful down the
road. The issue is that we can only get dependencies from
the source version of a class, so for an incremental compile
where we don't have the source type, this would resolve it
incrementally. This will pull in a lot of source unnecessarily
though so the fix was to save the syncTypeNames in the build info
so we don't have to recompute it.