do not use the sync filter for the remote method calls that originate
from the server. We usually use during the serialize to keep the client
from getting requests that are not bound for that particular client.
But for the simpleScope test script, we are targeting the output_c() method
call to the ScopeAppBar component that is in the appSession scope. We only
add the syncFilter to the currently targeted window, but there might be
other windows that should receive the method call. This change just sends
the remote method call to each client that is targeted by the scope object
to avoid that error. Another approach would be to store the syncTypeFilter
for the appSession context as well as the window, and search all appropriate
filters for a match. For the invokeRemoteMethod call, we could then add it
to the appSession's syncTypeFilter.