fixed bug that skipped transform for some glue statements
- In the problematic case, during transform of the glue statement
we'd call transformTemplateDeclaration() which would replace the
glue statement with a BlockStatement but where the block statement
would never get transformed.

just because transformTemplateDeclaration is used in some many places,
added a method to get the index of the current semantic node so that
we could find the replaced one and do the transform.