fixed bug with initializing data binding rules with db
We may not have a DBObject instance when creating the wrapper, and
even if we do, we may not have an id and so return null when accessing
a property that's persistent from the constructor. If then there's a
data binding rule that uses that value, it's now stale, waiting for the
setX method to set the value from the DB. But if this is a hollow instance
created from a reference, we will not have fetched that property and may not
fetch it. Unfortunately a data binding expression might have some other computed
value that's stale as well so when we get that (e.g. inStock in PhysicalSku) we
do not know to fetch the properties in the binding.

So to fix this problem, the fetchLater properties are registered when we are
deserializing a persistent object and are in the "applyingCommittedChanges" state
which should catch all of the cases where this is a problem. After we have retrieved
the properties, closed the result set, we just need to fetch any of these initial
stale properties so that we call the setX, send the events, and update the bindings.