fixed bug with tagMerge="replace"
- when getting a modified type that's a child of a parent that's
replaced, not modified a previous type, need to restrict the
search of the modified inner type to only the enclosing class.
otherwise, you find the modified type that's been replaced.

- for test scripts, when types are temporary, a modify inner type
of a class still needs to find the previous type in 'script mode'
where we don't create a modify for the parent class