changed model for how client-side page JS obejcts are created
previously, when generating the code for the top level class we'd
add code to create an instance using the getInstanceTemplate method.
For most components it's fine even though classes get instantiated
right when the code is loaded... that's because these are marked
as 'main objects' with @MainInit or one of the other ways to set that.

For page objects though, it means that all top-level classes get instantiated right
away, when the code for the of that class is loaded. This is kind of
limited and breaks when one top-level page class extends another.
Ideally, the base class is abstract so it is not registered with 'addPage'
and considered a top-level class anyway, but if you just want to customize an
existing top-level class, it requires refactoring just to give it a new path name
and customize the base class.

We might change this to include all main objects, but right it only works
for page objects. We added a call to js_PageInfo_c.initMatchingPage() that
uses the window.location.pathname to compare against the pageInfo's url pattern.
(currently only doing this if there is more than one page object).
We create the object instance and init it as a sync type if necessary.

After this, it will not be hard to detect URL changes, lazily create
the instance the first time, and support swapping from one page object
to another seamlessly, maybe for modal apps, or to support more complete
navigation scenarios declaratively. We'd need to create a module that
holds all of the page objects for all of the URLs. Then just navigate to
different URLs to choose the new page object.