bug fix for case with messed up getSet conversion
      totalResults += group.results.size();

when both totalResults, and group are converted, we were losing
the size() in the generated result.  It would get converted first to

 totalResults = totalResults + group.results.size()

and then into:

 totalResults = getTotalResults() + group.getResults().size()

where the last term is a new selector expression that needs to replace
the old identifierExpression in the BinaryOperand that is it's parent.
But we were not replacing it in the 'tree version' of the expression..
it would still be pointing to the identifier expression which has
had it's last term chopped off (since the identifier expression is
reused as the base part of the selector expression).

Fixed it by replacing those changes in the tree version but do not
trigger semantic node events since that part is not mapped onto the
parselet mapping.