fixed bug with session scoped state not being sync'd properly
We were losing a change when a property went from false to true
where the session scope was not updating its previous value of
the change. Later when it went back to false, it thought the value
was unchanged and the client did not receive the change.

also added a fix that seemed necessary, where we skip queuing
changes for these shared scopes since the change should always
be queued at the window level