fixes for merging annotations, to pick up JS dependent types
* added capability so that we can merge the values of annotations
rather than just overriding it. We need this so that we can pick up
all of the dependent types in a list of independent features.

* fixed a bug with how we were merging annotations - need to only
merge one-level of the modify-type at a time because we are only
guaranteed to have the transformedType for the immediate base type.
If we use the getAnnotation(.., true) - that checks the entire stack,
it will find the annotation further up the chain if it's there and
replace the value in the non-transformed mode, where it ultimately
gets lost.