turn sync onDemand off for objects by default
might need to add a way to override this, or maybe we should
get rid of this and just set the annotation. It's for product view
that needs to be on demand when it's create from a class using the
product view and not ondemand when it's instantiated in the ProductPage.