added sync init operation
once the server has invalidated the session, we also need
to tell the client that it should clear it's sync'd state.

This stops the 'auto sync' call but then when that window
makes its first request, we need to send it back the init
state again as part of the normal request. This fixes the login
/logout sequence and ensures one client doesn't get any objects
from the old stale login session