fixed security problem with 'class' in JSON
the JSON used for DB storage uses JSON 'class' as a way to
create a specific class. Since it's only set from trusted code
it's not vulnerable to misuse in creating random classes. But when
we use JSON in RPC, we do not allow this feature currently. We require
that all remote classes have sync'd types and so they will come across
the wire as regular references where there are security checks to be
sure we can create class.

As part of this change, I also ensure that the class created is a subclass
 of the expected property type. Even for the DB case, this will help restrict
what classes can be defined from these DB JSON entries.