don't set fromStatement when cloning for a different system
It's really used for debugging and things like that so that
the transformed statement can point back to the source.

It was a weird case where we were using the from statement
to add debug registeration but that was in a different system.
The bug showed up when the parse node on the different system
was cleared.