partial fix to problem with trying to refresh one page from the wrong context
this arose from a case where we were setting the user session.last modified property
in a scheduled thread when flushing the user session cache and it triggered a
refresh for a user manager page that happend to be looking at that session.
We actually did not want to update the last modified time in that instance and
so fixed that bug a different way but this will at least print an error do a better
job of setting a flag to indicate the page still needs to be refreshed.  Once we
have another test case for this, we can add a line of code in the sync servlet
to refresh the page and also add some code to wake up the sync servlet.  Or maybe
the use case should use the cross scope binding mechanism that already handles
that.