support cross-scope page refresh
normally we say that when a binding is used used from more than
one scope, you should use cross-scoped bindings so that we know
to queue the change event to the other context. But one category
of those we can handle automatically are page refreshes so this
code supports that.