changes to cache property for tag objects
we need to do caching for server tags in general but the serverTag
flag doesn't get set before can call getInnerHTML() and so it was
getting a startTagCacheValid =true but where startTagTxt was null and
that meant next time we render, we'd get no content for that tag.

right now, we call registerSyncInst in addServerTag before we call
addServerTags on the children that sets the serverTag flag.
But registerSyncInst will get all sync'd properties and that includes
innerHTML and startTagTxt so the serverTag flag is not set so we
were looking at cache enabled.

- I think that going forward we should set the serverTag falg before
we register the syncInst
- we need to always keep startTagTxt and innerHTML caches for server tags
so that we can tell when they change to update the client but when caching
is disabled on a sub-tag, need some way to search and find and refresh it
on each page request.