memory leaks and added cache expiration for db package
NOTE: new method: DBTypeDescriptor.invalidateExpiredItems() is not called
yet so no items will get expired. Need to add a timer that runs with a
configurable timeout, and/or when cache size exceeds a threshold.

fixed bug with Element.stop that could stop property references like ElementEditor.pageType
with new getStopSemanticProps() method in SemanticNode. When the Element is parsed is will
stop public properties since those are AST nodes. When the Element is a tagobject created at runtime
it will not.

updated README