moved processes/runtimes from static state into SystemContext
rather than using static data structures, store the system context
in the main LayeredSystem.  This is necessary for the Java to JS converter
form that wants to have completed independent LayeredSystem stacks, one for
each type of converter and that's indendent of the layered system that's running
the web app driving the form.